refactor: 도서관 재고 reader 설정 클래스 분리
