+++ src/test/java/com/bookspot/batch/stock/reader/StockReaderConfigTest.java
@@ -1,5 +1,6 @@
-package com.bookspot.batch.stock;
+package com.bookspot.batch.stock.reader;
 
+import com.bookspot.batch.stock.StockStepConfig;
 import com.bookspot.batch.stock.data.LibraryStockCsvData;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,8 +11,8 @@ import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class StockStepConfigTest {
-    StockStepConfig config = new StockStepConfig();
+class StockReaderConfigTest {
+    StockReaderConfig config = new StockReaderConfig();
 
     private static final String TARGET_PATH = "부평구립삼산도서관 장서 대출목록 (2024년 12월).csv";
     private static final Resource SAMPLE_RESOURCE = new PathMatchingResourcePatternResolver().getResource(TARGET_PATH);
